list refactor
