+++ common/src/main/java/pepjebs/mapatlases/client/screen/CartographyTableAtlasButton.java
@@ -2,6 +2,7 @@ package pepjebs.mapatlases.client.screen;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.VirtualOverride;
 import net.mehvahdjukaar.moonlight.api.resources.assets.LangBuilder;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
@@ -112,7 +113,7 @@ public class CartographyTableAtlasButton extends AbstractWidget {
         }
     }
 
-    //@Override
+    @VirtualOverride("neoforge")
     public void onClick(double mouseX, double mouseY, int button) {
         onClick(mouseX, mouseY);
     }
