+++ common/src/main/java/pepjebs/mapatlases/client/screen/DimensionBookmarkButton.java
@@ -1,6 +1,7 @@
 package pepjebs.mapatlases.client.screen;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.VirtualOverride;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Tooltip;
@@ -68,7 +69,7 @@ public class DimensionBookmarkButton extends BookmarkButton {
         parentScreen.selectDimension(dimension);
     }
 
-    //@Override
+    @VirtualOverride("neoforge")
     public void onClick(double mouseX, double mouseY, int button) {
         onClick(mouseX, mouseY);
     }
